requirements.txt
requests==2.32.5
pytest==9.0.2
jsonschema==4.26.0
python-dotenv==1.2.2
